Adds ability to create and deprovision apps and modify configs
I'm working towards filling in some of the functionality in this
API towards the goal of writing a program that can create a new
test environment for Connect.

This PR adds the ability to create and deprovision apps, and
to get and modify an app's current configuration (i.e.,
environment variables).